NNControl/server/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-session
flask-cors
flask-socketio
python-dotenv
redis
simple-websocket
pycryptodome
